feat: Add data.ts for notes annual chart
