+++ app/components/notes-annual-chart/NotesAnnualChartSection.tsx
@@ -8,15 +8,14 @@ import {
   STATUS_COLORS,
 } from "~/components/graph";
 
+import {
+  generateMockData,
+  MOCK_PERIOD_OPTIONS,
+  type MonthlyNoteData,
+} from "./data";
+
 /** 公開率の色（オレンジ） */
 const PUBLICATION_RATE_COLOR = "#ffa726";
-
-/** ダミー期間オプション（モックデータの期間に対応） */
-const MOCK_PERIOD_OPTIONS = [
-  { value: "2024/10-2025/09", label: "2024/10 〜 2025/09" },
-  { value: "2023/10-2024/09", label: "2023/10 〜 2024/09" },
-  { value: "2022/10-2023/09", label: "2022/10 〜 2023/09" },
-];
 
 /** 凡例アイテムのProps */
 type LegendItemProps = {
@@ -59,50 +58,6 @@ type SeriesVisibility = {
   evaluating: boolean;
   unpublished: boolean;
   publicationRate: boolean;
-};
-
-/** 月ごとのノートデータ */
-type MonthlyNoteData = {
-  /** 月ラベル（例: "2024/10"） */
-  month: string;
-  published: number;
-  evaluating: number;
-  unpublished: number;
-  publicationRate: number;
-};
-
-/** 12ヶ月分のモックデータを生成 */
-const generateMockData = (): MonthlyNoteData[] => {
-  const months = [
-    "2024/10",
-    "2024/11",
-    "2024/12",
-    "2025/01",
-    "2025/02",
-    "2025/03",
-    "2025/04",
-    "2025/05",
-    "2025/06",
-    "2025/07",
-    "2025/08",
-    "2025/09",
-  ];
-
-  return months.map((month) => {
-    const published = Math.floor(Math.random() * 800) + 600;
-    const evaluating = Math.floor(Math.random() * 500) + 300;
-    const unpublished = Math.floor(Math.random() * 400) + 200;
-    const total = published + evaluating + unpublished;
-    const publicationRate = Math.round((published / total) * 100);
-
-    return {
-      month,
-      published,
-      evaluating,
-      unpublished,
-      publicationRate,
-    };
-  });
 };
 
 type NotesAnnualChartSectionProps = {
